service: Skip NodePort service entries for remote target frontends
With Clustermesh, load-balancing NodePort traffic on the source node can
cause issues.

For example, socket load-balancing may route traffic incorrectly and
send it to a backend on local cluster rather than the remote cluster,
due to the use of wildcard IP lookups in the datapath that don't
preserve cluster knowledge. For more details, see:
https://github.com/cilium/cilium/issues/24692#issuecomment-1517583306

Another issue may arise when using the clustermesh-apiserver in
combination with KPR, tunneling, WireGuard and the Host Firewall. In
this case, the routing becomes asymmetric, using the native device for
one direction and tunneling for the other, and may be dropped if
WireGuard encryption starts on one side before the other has configured
it. As a result, Cilium may fail to establish the initial connection to
clustermesh-apiserver NodePort services. For details, see:
https://github.com/cilium/cilium/issues/31209

To work around such issues, skip load-balancing NodePort traffic at the
source when the target IP is for a node in a remote cluster, but just
forward it and let the remote node handle it.

Suggested-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>